# usdcop-trading-dashboard/docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: GlobalMarkets Terminal — FE↔BE Contract
  version: 1.0.0
  description: >
    Frontend-facing API (BFF) for the USDCOP multi-asset RL trading platform.
    See specs/frontend-backend-contract.md for prose, RBAC matrix and the
    view→endpoint map. This file is the authoritative machine-readable subset.
  contact: { name: Platform, url: "https://github.com/pspedro19/USDCOP-RL-Models" }
servers:
  - url: /            # same-origin BFF (browser → Next.js)
    description: Browser-facing BFF (relative; no cross-origin from the client)
...
security:
  - sessionCookie: []          # default: authenticated session required
paths:
...
  /api/execution/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Authenticate against SignalBridge; sets the session cookie.
      x-upstream: "signalbridge-api POST /api/auth/login  (+ GET /api/users/me for role)"
      security: [{}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, captchaToken]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
                captchaToken: { type: string, description: "From GET /api/captcha (public forms)" }
      responses:
        "200": { $ref: "#/components/responses/SessionOk" }
        "401": { $ref: "#/components/responses/Error401" }
        "429": { $ref: "#/components/responses/Error429" }
...
  /api/execution/auth/reset-password:
    post:
      tags: [auth]
      operationId: resetPassword
      summary: Trigger a password-reset email.
      x-upstream: "signalbridge-api POST /api/auth/reset-password"
      security: [{}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [email], properties: { email: { type: string, format: email } } }
      responses:
        "202": { description: Always 202 (no account enumeration) }
...
  /api/billing/me:
    get:
      tags: [billing]
      operationId: getEntitlements
      summary: Server-resolved EFFECTIVE entitlements for the logged-in user.
      description: Source of truth for gating (plan + per-asset add-ons + execution scope).
      responses:
        "200":
          description: Entitlements
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Ok"
                  - properties: { data: { $ref: "#/components/schemas/Entitlements" } }
        "401": { $ref: "#/components/responses/Error401" }
...
  /api/analysis/chat:
    post:
      tags: [analysis]
      operationId: analysisChat
      summary: Assistant — streamed answer grounded in the current week's analysis.
      x-upstream: "Anthropic Messages API (key server-side only)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                messages: { type: array, items: { type: object, properties: { role: { type: string }, content: { type: string } } } }
                asset: { type: string }
      responses:
        "200": { description: "text/event-stream (token stream)", content: { text/event-stream: { schema: { type: string } } } }
        "429": { $ref: "#/components/responses/Error429" }
...
  /api/execution/signal-bridge/kill-switch:
    get:
      tags: [signalbridge]
      operationId: getKillSwitch
      x-upstream: "signalbridge-api GET /api/signal-bridge/kill-switch/status"
      responses: { "200": { $ref: "#/components/responses/GenericOk" } }
    post:
      tags: [signalbridge]
      operationId: setKillSwitch
      summary: Flip the global kill-switch (halts copy-trading immediately).
      security: [{ sessionCookie: [admin] }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [active], properties: { active: { type: boolean }, reason: { type: string } } } } } }
      responses: { "200": { $ref: "#/components/responses/GenericOk" }, "403": { $ref: "#/components/responses/Error403" } }

  # ============================ ADMIN ===================================
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: >
        Browser→BFF auth. NextAuth session cookie (httpOnly, Secure, SameSite=Lax).
        The scope list on an operation (e.g. [admin]) documents the RBAC role the
        BFF re-checks in-handler — it is NOT an OAuth scope. 401 if absent, 403 if role denies.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC signature for provider webhooks (server-to-server; verified before processing).
...
  responses:
    SessionOk:
      description: Authenticated; session cookie set.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Ok"
              - properties: { data: { $ref: "#/components/schemas/Session" } }
    GenericOk:
      description: Success envelope.
      content: { application/json: { schema: { $ref: "#/components/schemas/Ok" } } }
...
    Error401: { description: Not authenticated, content: { application/json: { schema: { $ref: "#/components/schemas/Err" } } } }
    Error403: { description: Authenticated but role/entitlement denies, content: { application/json: { schema: { $ref: "#/components/schemas/Err" } } } }
...
    Error429:
      description: Rate limited.
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds until retry }
        RateLimit-Remaining: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Err" } } }
  schemas:
    Ok:
      type: object
      required: [ok, data]
      properties:
        ok: { type: boolean, enum: [true] }
        data: {}
        meta:
          type: object
          properties:
            nextCursor: { type: string, nullable: true }
            asOf: { type: string, format: date-time }
    Err:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, enum: [false] }
        error: { $ref: "#/components/schemas/Error" }
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: FORBIDDEN, description: Stable machine code (SCREAMING_SNAKE) }
        message: { type: string, description: Human-readable; safe to surface }
        status: { type: integer, example: 403 }
        details: { type: object, additionalProperties: true }
        traceId: { type: string, description: W3C trace id (correlate in Jaeger) }
        path: { type: string }
        timestamp: { type: string, format: date-time }
    Role: { type: string, enum: [admin, developer, subscriber, free] }
    PlanId: { type: string, enum: [free, signals, auto] }
    Session:
      type: object
      nullable: true
      properties:
        userId: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        plan: { $ref: "#/components/schemas/PlanId" }
        status: { type: string, enum: [active, pending, suspended] }
    Entitlements:
      type: object
      description: Server-resolved truth for gating. The client renders from this; it never decides access itself.
      properties:
        plan: { $ref: "#/components/schemas/PlanId" }
        permissions:
          type: array
          items: { type: string, example: signals:read }
        unlockedAssetIds: { type: array, items: { type: string } }
        executionScope: { type: string, enum: [none, self, global] }
...